dashboard_macroeconomia/requirements.txt
streamlit
lightweight-charts
yfinance
pandas_ta
ipeadatapy
python-bcb